split lib in more modules
